requirements.txt
twarc==2.4.3
spacy==2.3.5
transformers==4.0.1
torch==1.7.1
scikit-learn==0.23.2
pandas==1.3.3
wordsegment==1.3.1
emoji==1.5.0
python-dotenv==0.19.1
influxdb==5.3.1